Updated speed.js
Created global _events variable to hold event arrays for event management of on and off functions.

Got rid of the Object.assign and stored elements in an internal _elements variable.

Updated the each function to use the _elements variable instead of this.

Rewrote off for the most part, to allow deletion of all events, specific event types, and provided events for specified elements.